# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/users": {
            "get": {
                "description": "Get all users from the database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "List all users",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/user.User"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new user with the input payload",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Add a new user",
                "parameters": [
                    {
                        "description": "User info",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.CreateUserInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/user.User"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "server.ErrorResponse": {
            "type": "object",
            "required": [
                "code",
                "message"
            ],
            "properties": {
                "code": {
                    "description": "Code is a predefined error code string, e.g. \"0001\".",
                    "type": "string",
                    "example": "0000"
                },
                "message": {
                    "description": "Message should clearly describe the error in human-readable form.",
                    "type": "string",
                    "example": "The error in human-readable form"
                }
            }
        },
        "user.CreateUserInput": {
            "type": "object",
            "required": [
                "email",
                "name"
            ],
            "properties": {
                "email": {
                    "description": "Email address of the user, must be a valid email format",
                    "type": "string",
                    "example": "user@example.com"
                },
                "name": {
                    "description": "Name of the user",
                    "type": "string",
                    "minLength": 3,
                    "example": "John Doe"
                }
            }
        },
        "user.User": {
            "type": "object",
            "required": [
                "email",
                "name"
            ],
            "properties": {
                "email": {
                    "description": "Email of the user. Must be valid email format.",
                    "type": "string",
                    "example": "user@example.com"
                },
                "id": {
                    "description": "ID is the primary key, autogenerated by the database.",
                    "type": "integer"
                },
                "name": {
                    "description": "Name of the user.",
                    "type": "string",
                    "minLength": 3,
                    "example": "John Doe"
                }
            }
        }
    }
}