# internal/user/model.go
package user

// User represents a user record.
// Used in GET responses, includes the database ID.
type User struct {
	// ID is the primary key, autogenerated by the database.
	ID uint `json:"id" gorm:"primaryKey"`

	// Name of the user.
	Name string `json:"name" example:"John Doe" binding:"required,min=3"`

	// Email of the user. Must be valid email format.
	Email string `json:"email" example:"user@example.com" binding:"required,email"`
}

// CreateUserInput For Create requests (no ID)
type CreateUserInput struct {
	// Name of the user
	Name string `json:"name" example:"John Doe" binding:"required,min=3"`

	// Email address of the user, must be a valid email format
	Email string `json:"email" example:"user@example.com" binding:"required,email"`
}